If helperd hasn't checked in yet, don't send him a SIGHUP.
git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@20178 e92bca22-270c-0410-9cea-e3f1106b6a1c

+++ launchd/src/launchd.c
@@ -1582,7 +1582,7 @@ static void workaround3048875(int argc, char *argv[])
 
 static void notify_helperd(void)
 {
-	if (helperd && helperd->p)
+	if (helperd && helperd->p && helperd->checkedin)
 		kill(helperd->p, SIGHUP);
 }
